requirements.txt
confluent_kafka==2.9.0
Faker==37.1.0
pandas==2.2.3
prophet==1.1.6
psycopg2-binary
python-dotenv==1.1.0
streamlit==1.43.2
